[FE]: [REFACTOR] store 폴더 구조 변경
- actionTypes를 개별 스토어에서 전역 스토어로 이전
- ERROR, RESET 와 같이 여러 스토어에서 공통적으로 사용되는 액션 변경
