+++ client/src/components/molecules/UploadArea/UploadArea.tsx
@@ -4,8 +4,9 @@ import styled from 'styled-components';
 
 import Button from '@/components/atoms/Button';
 import FileInput from '@/components/atoms/FileInput';
-import { setVideo, reset } from '@/store/originalVideo/actions';
+import { setVideo } from '@/store/originalVideo/actions';
 import { getName } from '@/store/selectors';
+import { reset } from '@/store/actionTypes';
 
 const StyledDiv = styled.div`
   display: flex;
