--- client/src/store/originalVideo/actionTypes.ts
@@ -1,11 +1,0 @@
-export const FETCH_START = 'original/FETCH_START';
-export const SET_VIDEO = 'original/SET_VIDEO';
-export const LOAD_METADATA = 'original/LOAD_METADATA';
-export const LOAD_SUCCESS = 'original/LOAD_SUCCESS';
-/*
-export const PROCESS_START = 'original/PROCESS_START';
-export const PROCESS_SUCCESS = 'original/PROCESS_SUCCESS';
-export const PROCESS_ERROR = 'original/PROCESS_ERROR';
-*/
-export const RESET = 'original/RESET';
-export const ERROR = 'original/ERROR';
